# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/login:
    post:
      operationId: loginAdmin
      tags: [auth]
      summary: Log in as admin
      description: Verifies admin credentials and starts an admin session (sets a session cookie)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
...
components:
  schemas:
...
    AdminCredentials:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
        - username
        - password
    AdminSession:
      type: object
      properties:
        isAdmin:
          type: boolean
      required:
        - isAdmin
    AuthError:
      type: object
      properties:
        error:
          type: string
      required:
        - error